pyproject.toml
[project]
name = "kizilelma"
version = "0.1.0"
description = "Türkiye finansal piyasa enstrümanlarını izleyen kişisel yatırım danışmanı ajanı"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "feedparser>=6.0.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
    "anthropic>=0.25.0",
    "apscheduler>=3.10.0",
    "sqlmodel>=0.0.16",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[project.scripts]
kizilelma = "kizilelma.main:cli"
kizilelma-web = "kizilelma.web.app:main"
